+++ example/lib/main.dart
@@ -18,7 +18,6 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     restoreWallet(
-      //"curtain stairs horse sing file upset laundry cherry make oblige ticket medal",
         "puppy interest whip tonight dad never sudden response push zone pig patch",
         Network.TESTNET);
     super.initState();
@@ -36,16 +35,10 @@ class _MyAppState extends State<MyApp> {
     print("mnemonicWithEntropy: $mnemonicWithEntropy5 ");
   }
 
-
   restoreWallet(String mnemonic, Network network) async {
+    final res = await createDescriptors(type: Descriptor.P2PK, mnemonic:  mnemonic, network: network );
     final resWallet = await  bdkFlutter.createWallet(
-        mnemonic: mnemonic,
-        network: network,
-        blockchainConfig: BlockchainConfig.electrum(
-            config: ElectrumConfig(
-                stopGap: 10,
-                retry: 5,
-                url: "ssl://electrum.blockstream.info:60002"))
+       descriptor: res.descriptor
     );
     print(resWallet.address);
     print(resWallet.balance.total);
@@ -127,7 +120,7 @@ getConfirmedTransactions() async {
                   child: const Text('Press to create new Address')),
               TextButton(
                   onPressed: () => sendBit(),
-                  child: const Text('Press to  send 1200 satoshi')),
+                  child: const Text('Press to  send 1200 sats')),
               TextButton(
                   onPressed: () => sync(), child: const Text('Press to  sync')),
               TextButton(
@@ -135,7 +128,7 @@ getConfirmedTransactions() async {
                   child: const Text('Get ConfirmedTransactions')),
               TextButton(
                   onPressed: () => getPendingTransactions(),
-                  child: const Text('getPendingTransactions')),
+                  child: const Text('get PendingTransactions')),
               TextButton(
                   onPressed: () => getBalance(),
                   child: const Text('get Balance')),