+++ lib/src/utils/exceptions/broadcast_exceptions.dart
@@ -3,7 +3,9 @@ part 'broadcast_exceptions.freezed.dart';
 
 @freezed
 class BroadcastException with _$BroadcastException {
-  const factory BroadcastException.insufficientFunds(String e) =  InsufficientFunds;
-  const factory BroadcastException.insufficientBroadcastAmount(String e) =  InsufficientBroadcastAmount;
+  const factory BroadcastException.insufficientFunds(String e) =
+      InsufficientFunds;
+  const factory BroadcastException.insufficientBroadcastAmount(String e) =
+      InsufficientBroadcastAmount;
   const factory BroadcastException.unexpected(String e) = Unexpected;
 }