+++ CHANGELOG.md
@@ -1,3 +1,20 @@
-## 0.0.1
+## [0.1.4]
 
-* TODO: Describe initial release.
+#### Functionality Added
+- Generate Mnemonic method
+- Create Descriptors
+- Create Extended Key
+- Create XPriv
+- Create XPub
+- Create Wallet
+- Get New Address
+- Get Last Unused Address
+- Get Balance
+- Get Transactions
+- Get Pending Transactions
+- Get Confirmed Transactions
+- Sync Wallet
+- Create Transaction
+- Sign Transaction
+- Broadcast Transaction
+- Quick Send